requirements.txt
# If your environment already runs Qwen2SAM_Detecture, its deps also cover
# this project — you can skip `pip install -r requirements.txt`. This file
# is for fresh setups only.

# Core ML (known-working here: torch 2.6.0+cu124, transformers 5.1.0).
torch>=2.3
transformers>=4.56

# Qwen3-VL tokenizer/image utils
# qwen-vl-utils.process_vision_info is REQUIRED — on transformers >= 5.x the
# bare `{"type": "image"}` placeholder pattern desyncs the image grid and
# crashes CUDA with a gather-kernel OOB.
qwen-vl-utils
sentencepiece
tiktoken

# Vision / IO
opencv-python-headless
Pillow
# SAM3 + dctorch require numpy<2. tiktoken's install otherwise pulls
# numpy 2.x transitively.
numpy<2,>=1.26

# Matching + metrics
scipy

# Visualization
matplotlib
